custom-drawers: Works with complex fields. Need some optimizations and refactoring.
