+++ Assets/OneLine/OneLine/Example/Details/Custom Property Drawer Example.asset.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 9d49831d8c35945b59ed870fb434e49a
-timeCreated: 1519298473
+guid: 4da82b0dc13594f77941459b270cafac
+timeCreated: 1519373945
 licenseType: Free
 NativeFormatImporter:
   mainObjectFileID: 11400000
