+++ Assets/OneLine/OneLine/Example/Details/Scripts/CustomPropertyDrawerExample.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: df9cda7dbd5e943a38c44d0a82e6cd6e
+timeCreated: 1519373913
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
